CLI: added list instances and plan status commands + fix of status of wrong instance
