+++ cmd/maestroctl/cmd/plan.go
@@ -8,12 +8,11 @@ import (
 func NewPlanCmd() *cobra.Command {
 	newCmd := &cobra.Command{
 		Use:   "plan",
-		Short: "-> Show all available plans.",
-		Long:  `The plan command has subcommands to show all available plans.`,
+		Short: "-> View all available plans.",
+		Long:  `The plan command has subcommands to view all available plans.`,
 	}
 
 	newCmd.AddCommand(plan.NewPlanHistoryCmd())
-	newCmd.AddCommand(plan.NewPlanListCmd())
 	newCmd.AddCommand(plan.NewPlanStatusCmd())
 
 	return newCmd
